+++ crates/ruma-events-macros/src/event_enum.rs
@@ -89,7 +89,7 @@ fn expand_event_enum(
             )*
             /// An event not defined by the Matrix specification
             #[doc(hidden)]
-            _Custom(#ruma_events::#event_struct<#ruma_events::custom::_CustomEventContent>),
+            _Custom(#ruma_events::#event_struct<#ruma_events::_custom::CustomEventContent>),
         }
 
         #deserialize_impl