+++ crates/ruma-signatures/src/lib.rs
@@ -46,7 +46,7 @@
 
 use ruma_serde::{AsRefStr, DisplayAsRefStr};
 
-pub use error::{Error, JsonError, JsonType, ParseError, SplitError, VerificationError};
+pub use error::{Error, JsonError, JsonType, ParseError, VerificationError};
 pub use functions::{
     canonical_json, content_hash, hash_and_sign_event, redact, redact_content_in_place,
     redact_in_place, reference_hash, sign_json, verify_event, verify_json,
@@ -72,7 +72,7 @@ pub enum Algorithm {
 }
 
 /// Extract the algorithm and version from a key identifier.
-fn split_id(id: &str) -> Result<(Algorithm, String), SplitError> {
+fn split_id(id: &str) -> Result<(Algorithm, String), Error> {
     /// The length of a valid signature ID.
     const SIGNATURE_ID_LENGTH: usize = 2;
 
@@ -81,20 +81,20 @@ fn split_id(id: &str) -> Result<(Algorithm, String), SplitError> {
     let signature_id_length = signature_id.len();
 
     if signature_id_length != SIGNATURE_ID_LENGTH {
-        return Err(SplitError::InvalidLength(signature_id_length));
+        return Err(Error::InvalidLength(signature_id_length));
     }
 
     let version = signature_id[1];
 
     if !version.bytes().all(|ch| ch.is_ascii_alphanumeric() || ch == b'_') {
-        return Err(SplitError::InvalidVersion(version.into()));
+        return Err(Error::InvalidVersion(version.into()));
     }
 
     let algorithm_input = signature_id[0];
 
     let algorithm = match algorithm_input {
         "ed25519" => Algorithm::Ed25519,
-        algorithm => return Err(SplitError::UnsupportedAlgorithm(algorithm.into())),
+        algorithm => return Err(Error::UnsupportedAlgorithm(algorithm.into())),
     };
 
     Ok((algorithm, signature_id[1].to_owned()))